+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * [#654](https://github.com/bbatsov/rubocop/pull/654): `For` cop is now configurable to enforce either `each` (default) or `for`. ([@jonas054][])
 * Allow to overwrite the severity of a cop with the new `Severity` param. ([@codez][])
 * New cop `FlipFlop` checks for flip flops. ([@agrimm][])
+* [#577](https://github.com/bbatsov/rubocop/issues/577): Introduced `MethodDefParentheses` to allow for for requiring either parentheses or no parentheses in method definitions. Replaces `DefWithoutParentheses`. ([@skanev][])
 
 ### Changes
 
@@ -603,3 +604,4 @@
 [@chulkilee]: https://github.com/chulkilee
 [@codez]: https://github.com/codez
 [@emou]: https://github.com/emou
+[@skanev]: http://github.com/skanev